Refactor Forth
Replace separate ROW_[A-C] words with CUBE_ROW_PART defined at top

Combine PART_2[AB] into PART_2, and create CUBE_ROW at top of file to
use for it

For each word defined, make sure that it either ends with the stack in
the same stat as it begins, or consume exactly 1 value from the stack,
and document which they do. Each part leaves the stack exactly as it
was.

Use consistent commenting style except for SPDX header, as reuse doesn't
detect it properly within parentheses