Scheme refactor
One of the least significant changes - I was already using a similar
approach to the one I'm trying to use in the refactorings, but had
separate row_[a-c] procs, and defined them within the color_cube code.

Now, the procedure names are changed for consistenct with other
implementations, and row_[a-c] procs are replaced with a cube_row_part
proc, which creates a range and maps it to color_cell, because that was
done every time a row_[a-c] proc was called anyway.